Recover the true parent id for primary particles
Previously for primary particles the parent id was set to -1 when constructing the TParticle.
This would not allow to re-construct the hierarchy of the generated particles at a later stage.
In case a valid secondparentID is found (as passed by FairPrimaryGenerator) the parentId is overridden.
The behaviour for VMC particles is unmodified as secondparentID = -1 in this cases.